# Double buffer metrics In Train to avoid gap between steps